+++ swift/Sources/NaughtBotE2EEPayloads/Types.swift
@@ -1108,27 +1108,27 @@ public enum Components {
             ///
             /// - Remark: Generated from `#/components/schemas/MailboxSshAuthResponseSuccessV1/signature`.
             public var signature: OpenAPIRuntime.Base64EncodedData
-            /// Per-signature SK assertion flags byte the signer's secure element actually asserted with. MAY differ from the request `flags` byte (e.g. when the SK could not deliver user verification it MUST clear `0x04` on the assertion). Receivers MUST embed this byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead.
+            /// Per-signature SK assertion flags byte the signer's secure element actually asserted with. Approvers MUST either (a) assert with at least the bits the request `flags` byte asked for (UP=0x01, UV=0x04) and return the resulting byte here, or (b) return a `MailboxSshAuthResponseFailureV1` / `MailboxSshSignResponseFailureV1` with the appropriate signing error code. Approvers MUST NOT return a success response whose asserted flags byte clears bits the requester set; that would silently downgrade the security posture (e.g. UV-required → UP-only) below what the request agreed to. Receivers MUST embed this asserted byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead. Receivers SHOULD additionally verify that every bit set in the request `flags` byte is also set here as belt-and-suspenders defence against a misbehaving approver.
             ///
             /// - Remark: Generated from `#/components/schemas/MailboxSshAuthResponseSuccessV1/flags`.
             public var flags: Swift.Int
-            /// Monotonic counter the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)`. Successive signatures from the same key handle MUST have strictly increasing counter values.
+            /// Monotonic counter (u32) the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)` as a 4-byte big-endian unsigned integer. Successive signatures from the same key handle MUST have strictly increasing counter values. The schema declares `format: int64` so 32-bit Go targets can still represent the full u32 range without overflow.
             ///
             /// - Remark: Generated from `#/components/schemas/MailboxSshAuthResponseSuccessV1/counter`.
-            public var counter: Swift.Int
+            public var counter: Swift.Int64
             /// - Remark: Generated from `#/components/schemas/MailboxSshAuthResponseSuccessV1/approval_proof`.
             public var approval_proof: Components.Schemas.ApprovalAttestedKeyProof?
             /// Creates a new `MailboxSshAuthResponseSuccessV1`.
             ///
             /// - Parameters:
             ///   - signature: RFC 4648 standard base64 with `=` padding for the raw SSH signature blob (no SSH-wire framing).
-            ///   - flags: Per-signature SK assertion flags byte the signer's secure element actually asserted with. MAY differ from the request `flags` byte (e.g. when the SK could not deliver user verification it MUST clear `0x04` on the assertion). Receivers MUST embed this byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead.
-            ///   - counter: Monotonic counter the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)`. Successive signatures from the same key handle MUST have strictly increasing counter values.
+            ///   - flags: Per-signature SK assertion flags byte the signer's secure element actually asserted with. Approvers MUST either (a) assert with at least the bits the request `flags` byte asked for (UP=0x01, UV=0x04) and return the resulting byte here, or (b) return a `MailboxSshAuthResponseFailureV1` / `MailboxSshSignResponseFailureV1` with the appropriate signing error code. Approvers MUST NOT return a success response whose asserted flags byte clears bits the requester set; that would silently downgrade the security posture (e.g. UV-required → UP-only) below what the request agreed to. Receivers MUST embed this asserted byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead. Receivers SHOULD additionally verify that every bit set in the request `flags` byte is also set here as belt-and-suspenders defence against a misbehaving approver.
+            ///   - counter: Monotonic counter (u32) the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)` as a 4-byte big-endian unsigned integer. Successive signatures from the same key handle MUST have strictly increasing counter values. The schema declares `format: int64` so 32-bit Go targets can still represent the full u32 range without overflow.
             ///   - approval_proof:
             public init(
                 signature: OpenAPIRuntime.Base64EncodedData,
                 flags: Swift.Int,
-                counter: Swift.Int,
+                counter: Swift.Int64,
                 approval_proof: Components.Schemas.ApprovalAttestedKeyProof? = nil
             ) {
                 self.signature = signature
@@ -1153,7 +1153,7 @@ public enum Components {
                     forKey: .flags
                 )
                 self.counter = try container.decode(
-                    Swift.Int.self,
+                    Swift.Int64.self,
                     forKey: .counter
                 )
                 self.approval_proof = try container.decodeIfPresent(
@@ -1371,27 +1371,27 @@ public enum Components {
             ///
             /// - Remark: Generated from `#/components/schemas/MailboxSshSignResponseSuccessV1/signature`.
             public var signature: OpenAPIRuntime.Base64EncodedData
-            /// Per-signature SK assertion flags byte the signer's secure element actually asserted with. MAY differ from the request `flags` byte (e.g. when the SK could not deliver user verification it MUST clear `0x04` on the assertion). Receivers MUST embed this byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead.
+            /// Per-signature SK assertion flags byte the signer's secure element actually asserted with. Approvers MUST either (a) assert with at least the bits the request `flags` byte asked for (UP=0x01, UV=0x04) and return the resulting byte here, or (b) return a `MailboxSshAuthResponseFailureV1` / `MailboxSshSignResponseFailureV1` with the appropriate signing error code. Approvers MUST NOT return a success response whose asserted flags byte clears bits the requester set; that would silently downgrade the security posture (e.g. UV-required → UP-only) below what the request agreed to. Receivers MUST embed this asserted byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead. Receivers SHOULD additionally verify that every bit set in the request `flags` byte is also set here as belt-and-suspenders defence against a misbehaving approver.
             ///
             /// - Remark: Generated from `#/components/schemas/MailboxSshSignResponseSuccessV1/flags`.
             public var flags: Swift.Int
-            /// Monotonic counter the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)`. Successive signatures from the same key handle MUST have strictly increasing counter values.
+            /// Monotonic counter (u32) the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)` as a 4-byte big-endian unsigned integer. Successive signatures from the same key handle MUST have strictly increasing counter values. The schema declares `format: int64` so 32-bit Go targets can still represent the full u32 range without overflow.
             ///
             /// - Remark: Generated from `#/components/schemas/MailboxSshSignResponseSuccessV1/counter`.
-            public var counter: Swift.Int
+            public var counter: Swift.Int64
             /// - Remark: Generated from `#/components/schemas/MailboxSshSignResponseSuccessV1/approval_proof`.
             public var approval_proof: Components.Schemas.ApprovalAttestedKeyProof?
             /// Creates a new `MailboxSshSignResponseSuccessV1`.
             ///
             /// - Parameters:
             ///   - signature: RFC 4648 standard base64 with `=` padding for the raw SSH signature blob (no SSH-wire framing).
-            ///   - flags: Per-signature SK assertion flags byte the signer's secure element actually asserted with. MAY differ from the request `flags` byte (e.g. when the SK could not deliver user verification it MUST clear `0x04` on the assertion). Receivers MUST embed this byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead.
-            ///   - counter: Monotonic counter the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)`. Successive signatures from the same key handle MUST have strictly increasing counter values.
+            ///   - flags: Per-signature SK assertion flags byte the signer's secure element actually asserted with. Approvers MUST either (a) assert with at least the bits the request `flags` byte asked for (UP=0x01, UV=0x04) and return the resulting byte here, or (b) return a `MailboxSshAuthResponseFailureV1` / `MailboxSshSignResponseFailureV1` with the appropriate signing error code. Approvers MUST NOT return a success response whose asserted flags byte clears bits the requester set; that would silently downgrade the security posture (e.g. UV-required → UP-only) below what the request agreed to. Receivers MUST embed this asserted byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead. Receivers SHOULD additionally verify that every bit set in the request `flags` byte is also set here as belt-and-suspenders defence against a misbehaving approver.
+            ///   - counter: Monotonic counter (u32) the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)` as a 4-byte big-endian unsigned integer. Successive signatures from the same key handle MUST have strictly increasing counter values. The schema declares `format: int64` so 32-bit Go targets can still represent the full u32 range without overflow.
             ///   - approval_proof:
             public init(
                 signature: OpenAPIRuntime.Base64EncodedData,
                 flags: Swift.Int,
-                counter: Swift.Int,
+                counter: Swift.Int64,
                 approval_proof: Components.Schemas.ApprovalAttestedKeyProof? = nil
             ) {
                 self.signature = signature
@@ -1416,7 +1416,7 @@ public enum Components {
                     forKey: .flags
                 )
                 self.counter = try container.decode(
-                    Swift.Int.self,
+                    Swift.Int64.self,
                     forKey: .counter
                 )
                 self.approval_proof = try container.decodeIfPresent(
